pyproject.toml
[project]
name = "ggj"
version = "0.1.0"
dependencies = [
    "pygame",
    "mypy",
    "pillow",
    "numpy"
]
scripts = { ggj = "ggj.main:main" }

[project.optional-dependencies]
dev = [
    "ruff",
    "pyright",
]

[tool.mypy]
check_untyped_defs = true